% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{append_a4_vertical}
\alias{append_a4_vertical}
\title{Append two A4 generated images vertically}
\usage{
append_a4_vertical(files, output)
}
\arguments{
\item{files}{Vector of file paths, without the 'A4_' prefix}

\item{output}{Output file path}
}
\description{
Designed for join two images generated by \code{save_png_a4} that are the full page width
}
