% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{Austral_season_month}
\alias{Austral_season_month}
\title{Get Austral season month of a date}
\usage{
Austral_season_month(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
A number from 1 to 12
}
\description{
Get Austral season month of a date
}
\examples{
Austral_season_month(lubridate::make_date(2012, 11, 2))
}
