% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{build_data_table_by_region}
\alias{build_data_table_by_region}
\title{Prepare a region based data table for processing}
\usage{
build_data_table_by_region(
  time_df,
  batch_data,
  target_variables,
  region_info,
  convert_temps_from_K_to_C = TRUE,
  convert_pr_from_rate_to_mm = TRUE,
  convert_epanave_from_energy_to_mm = TRUE
)
}
\arguments{
\item{time_df}{Data frame with time series data, e.g. produced from \code{make_time_dataframe_from_nc_files}}

\item{batch_data}{Spatial cell extract data for a group of cells as a list of values for each variable}

\item{target_variables}{Vector of variables to extract from the \code{batch_data} list}

\item{region_info}{Item from the target cells lookup list corresponding to the region being processed. Required for a method of 'by_region'}

\item{convert_temps_from_K_to_C}{If TRUE, will assume all temperature values are in K and will convert them to C (default: FALSE)}

\item{convert_pr_from_rate_to_mm}{If TRUE, will assume pr values are rate and will convert them to mm (default: TRUE)}

\item{convert_epanave_from_energy_to_mm}{If TRUE, will assume epanave values are in energy and will convert them to mm (default: TRUE)}
}
\value{
A data frame with the contents of \code{time_df} as the data for the with the contents of \code{batch_data} grouped into a single figure, with a column added designating the region
}
\description{
Prepare a region based data table for processing
}
