% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{build_multi_model_mean_for_polygon}
\alias{build_multi_model_mean_for_polygon}
\title{Build multi-model mean dataframe}
\usage{
build_multi_model_mean_for_polygon(
  source_data_directory,
  target_region_code,
  target_models = NULL,
  required_filename_part = NULL
)
}
\arguments{
\item{source_data_directory}{Directory with saved model}

\item{target_region_code}{Region to build dataframe for}

\item{target_models}{A vector of models to include. If NULL, then all models for data in the directory will be included.}

\item{required_filename_part}{Permits further filtering of which model files to include. Only files that contain this text will be used.}
}
\value{

}
\description{
Build multi-model mean dataframe
}
