% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{build_multi_model_mean_from_mme}
\alias{build_multi_model_mean_from_mme}
\title{Build multi-model mean dataframe from pre-existing multi-model ensemble}
\usage{
build_multi_model_mean_from_mme(df_mme, group_by_vars = NULL)
}
\arguments{
\item{df_mme}{Dataframe of a multi-model ensemble}
}
\value{

}
\description{
Build multi-model mean dataframe from pre-existing multi-model ensemble
}
