% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{collate_polygon_data_for_model}
\alias{collate_polygon_data_for_model}
\title{Collate data from an extract directory for a given polygon for a given model}
\usage{
collate_polygon_data_for_model(
  extract_directory,
  target_cells_lookup,
  target_region_code,
  target_model
)
}
\arguments{
\item{extract_directory}{Directory with source data files named like "{1}_{2}_{3}" where: {1} is the GCM name, {2} is the domain code, {3} is the range of cell numbers in that file}

\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{target_region_code}{Region code to be added as a column to this dataframe}

\item{target_model}{Model to use}
}
\value{

}
\description{
Collate data from an extract directory for a given polygon for a given model
}
