% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{colour_lightener}
\alias{colour_lightener}
\alias{colour_darkener}
\title{Lighten or darken colours}
\usage{
colour_lightener(cols, x = 0.5)

colour_darkener(cols, x = 0.5)
}
\arguments{
\item{cols}{List or vector of colours}

\item{x}{Factor to adjust colour by. A value of 1 will return the original colour;
for \code{colour_lightener} a value of 0 will return white,
and for \code{colour_darkener} a value of 0 will return black,}
}
\value{

}
\description{
desc
}
\details{
details
}
