% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\docType{data}
\name{common_defaults}
\alias{common_defaults}
\title{Common default values}
\format{
An object of class \code{list} of length 1.
}
\usage{
common_defaults
}
\description{
A placeholder to contain default values for parameters that are used in multiple functions
}
\keyword{internal}
