% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{convert_all_polygons_to_multipolygons}
\alias{convert_all_polygons_to_multipolygons}
\title{Convert all geometries of a simple feature collection to a type of MULTIPOLYGON}
\usage{
convert_all_polygons_to_multipolygons(
  polygons_all_sf,
  field_names = list(index = "OBJECTID", label = "LABEL", code = "CODE")
)
}
\arguments{
\item{polygons_all_sf}{A simple feature collection}

\item{field_names}{A named list with the following names: 'index', 'label', and 'code'}
}
\value{
A list of simple features
}
\description{
Converts a simple feature collection, where each feature may have multiple polygons making up its geometry,
to a list of simple features only having one MULTIPOLYGON object making up its geometry.
}
\details{
The field list points to the columns in the simple feature that refer to:
* index: the unique index (usually OBJECTID)
* label: the label for each feature

And to be generated, *TODO if it does not exist TODO*:
* code: a unique code for each feature (this will be derived from the label)

In most cases, the default field names of ('OBJECTID', 'CODE', 'LABEL') will be incorrect
}
