% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{convert_cells_to_lat_lon_polys}
\alias{convert_cells_to_lat_lon_polys}
\title{Convert cell orography information to spatial polygons}
\usage{
convert_cells_to_lat_lon_polys(
  plot_cell_points,
  proj4string = common_defaults$proj4string
)
}
\arguments{
\item{plot_cell_points}{Dataframe of cells to return polygons for, taken from a cell orography lookup}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}
}
\value{
SpatialPolygons object with polygons for all of the cells
}
\description{
Takes a list of cells from a cell orography lookup and returns a list of SpatialPolygons of the dimensions of each cell
}
\details{
Useful to overlay cell boundaries on a plot.
}
