% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_cell_orography_lookup_from_domain_list}
\alias{create_cell_orography_lookup_from_domain_list}
\title{Create cell orography lookup for a list of domains}
\usage{
create_cell_orography_lookup_from_domain_list(target_domains_files)
}
\arguments{
\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}
}
\value{
Dataframe with orography data as from \code{create_cell_orography_lookup}
}
\description{
Takes a named list of domains with example NetCDF files and passes each to \code{create_cell_orography_lookup} and binds the entire collection to a single dataframe
}
\examples{
target_domains_files <- list(
  'ABC-10' = "/path/to/file/ABC-10.nc",
  'XYZ-05' = "/path/to/file/XYZ-05.nc",
  )
create_cell_orography_lookup_from_domain_list(target_domains_files)
}
