% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_date_breaks_yearly}
\alias{create_date_breaks_yearly}
\title{Title}
\usage{
create_date_breaks_yearly(dates)
}
\arguments{
\item{dates}{Vector of dates (POSIXct or date)}
}
\value{
Vector of dates
}
\description{
for scale_x_datetime
}
\examples{
create_date_breaks_yearly(seq.Date(from = as.Date('1950-06-01'),
                                   to = as.Date('2020-06-01'),
                                   by = 'day'))
}
