% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_polygons_target_cells_lookup_from_multiple_shape_files}
\alias{create_polygons_target_cells_lookup_from_multiple_shape_files}
\title{Title}
\usage{
create_polygons_target_cells_lookup_from_multiple_shape_files(
  polygons_properties_table,
  target_domains_files,
  preferred_domains
)
}
\arguments{
\item{polygons_properties_table}{A dataframe specifying the shape files for each polygon, key fields for each polygon.
Required columns: filename (full path to file), index (), label, (), code ()}

\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}
}
\value{
A target cells lookup list
}
\description{
For a shape files from a mixture of sources, therefore a lot more must be specified
}
