% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_polygons_target_cells_lookup_from_sf}
\alias{create_polygons_target_cells_lookup_from_sf}
\title{Create a target cells lookup from an sf object}
\usage{
create_polygons_target_cells_lookup_from_sf(
  polygons_all_sf,
  target_domains_files,
  field_names = list(index = "OBJECTID", label = "LABEL", code = "CODE"),
  preferred_domains = list()
)
}
\arguments{
\item{polygons_all_sf}{Simple feature object with polygons for all regions}

\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}

\item{field_names}{A named list with the following names: 'index', 'label', and 'code'}
}
\value{
A target cells lookup list
}
\description{
For a single shape file, therefore a lot more can be assumed about the properties
}
