% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_polygons_target_cells_lookup_from_single_shape_file}
\alias{create_polygons_target_cells_lookup_from_single_shape_file}
\alias{create_polygons_target_cells_lookup_from_multiple_shape_files}
\title{Create a target cells lookup from shape files}
\usage{
create_polygons_target_cells_lookup_from_single_shape_file(
  all_polygons_shape_file,
  target_domains,
  cell_orog_lookup,
  field_names = list(index = "OBJECTID", label = "LABEL", code = "CODE"),
  preferred_domains = list()
)

create_polygons_target_cells_lookup_from_multiple_shape_files(
  polygons_properties_table,
  target_domains,
  cell_orog_lookup,
  preferred_domains = list()
)
}
\arguments{
\item{all_polygons_shape_file}{Shape file will polygons for all regions}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{field_names}{A named list with the following names: 'index', 'label', and 'code'}

\item{preferred_domains}{Optional. A named list defining the preferred domain for each polygon.
If the polygons are spread across multiple domains, then the preferred domain for each polygon must be specified explictly to distinguish between those that overlap.
Use the code for the polygon (as will be created in \code{convert_all_polygons_to_multipolygons}) as the key, and the doman as the value.
If only one domain is present in the \code{target_domains_files} parameter, then this list is not required.}

\item{polygons_properties_table}{A data frame specifying the shape files for each polygon, key fields for each polygon.
Required columns: filename (full path to file), index (), label, (), code ()}
}
\value{
A target cells lookup list
}
\description{
Details
}
\details{
For a single shape file, some information can be assumed about the properties,
whereas for multiple shape files from a mixture of sources, a lot more must be specified explicitly.
}
