% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_shapes_from_dataframe_of_points}
\alias{create_shapes_from_dataframe_of_points}
\title{Create sf points from lat/lon values}
\usage{
create_shapes_from_dataframe_of_points(
  dataframe_locations,
  dist = 0.05,
  proj4string = common_defaults$proj4string
)
}
\arguments{
\item{dataframe_locations}{Dataframe specifying locations.}

\item{dist}{numeric; buffer distance for all, or for each of the elements in \code{x}; in case
\code{dist} is a \code{units} object, it should be convertible to \code{arc_degree} if
\code{x} has geographic coordinates, and to \code{st_crs(x)$units} otherwise}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}
}
\value{
Simple feature collection
}
\description{
Creates a simple feature collection of sf buffers centered around points defined in a dataframe
}
\details{
Dataframe must contain latitidues, longitudes, and labels of the points;
in columns named 'lat', 'lon', and 'label', respectively.

The default buffer distance is 0.05
}
\examples{
dataframe_locations <- tribble(~lat, ~lon, ~label,
48.858222, 2.2945, 'Eiffel Tower',
51.5007, -0.1245, 'Big Ben')
x <- create_shapes_from_dataframe_of_points(dataframe_locations)
ggplot(x) + borders('world') + geom_sf(fill = 'red', colour = 'red') +
  coord_sf(xlim = c(-10, 10), ylim = c(45, 55))
}
