% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_time_series_dataframe}
\alias{create_time_series_dataframe}
\title{Create time series dataframe from a time series vector}
\usage{
create_time_series_dataframe(time_series, add_Austral_columns = TRUE)
}
\arguments{
\item{time_series}{Vector of POSIX dates}

\item{add_Austral_columns}{Add columns for southern hemisphere year, month, and day of season (aka day of Austral year)}
}
\value{
Dataframe with supplied dates and additional date related columns
}
\description{
From a vector of POSIX dates, create a dataframe with additional columns derived from the supplied dates.
}
