% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{date_from_year_month}
\alias{date_from_year_month}
\title{Convert a year and month to a POSIX date corresponding to the first of that month}
\usage{
date_from_year_month(year, month, tz = "AEDT")
}
\arguments{
\item{year}{Year}

\item{month}{Month}

\item{tz}{Timezone, as is passed to as.POSIXct (default: 'AEDT')}
}
\value{
POSIX date
}
\description{
Convert a year and month to a POSIX date corresponding to the first of that month
}
\examples{
date_from_year_month(2016, 7)
date_from_year_month(2016, 7, tz = 'UTC')
}
