% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{draw_spatial_coverage_plot}
\alias{draw_spatial_coverage_plot}
\title{Draw a spatial coverage plot}
\usage{
draw_spatial_coverage_plot(
  target_cells_lookup,
  plot_cell_points,
  region,
  domain,
  bg_map_sf = NULL,
  show_cell_labels = FALSE,
  proj4string = common_defaults$proj4string
)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{plot_cell_points}{Data frame with \code{lon} and \code{lat} values for each \code{cell_number} to plot}

\item{region}{Region (will be shown on the plot)}

\item{domain}{Domain (will be shown on the plot)}

\item{bg_map_sf}{An optional simple feature object to plot as a background map}

\item{show_cell_labels}{logical. If TRUE then shows the labels for the cell numbers. Set to FALSE by default.}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}
}
\value{
ggplot object for the plot
}
\description{
Draw a spatial coverage plot
}
\examples{
target_cells_lookup <- readRDS('/path/to/lookup_file.rds')
cell_orog_lookup <- readRDS('/path/to/orog_file.rds')
region_code <- 'glenrowan'
domain_code <- 'VIC-5'

plot_cell_points <- get_cells_from_orog_lookup_for_region_domain(target_cells_lookup, cell_orog_lookup, region = region_code, domain = domain_code)

draw_spatial_coverage_plot(target_cells_lookup, plot_cell_points, region = region_code, domain = domain_code, show_cell_labels = TRUE)
}
