% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{draw_spatial_coverage_plot_for_region_in_domain}
\alias{draw_spatial_coverage_plot_for_region_in_domain}
\title{Draw a spatial coverage plot for a region on a specific domain}
\usage{
draw_spatial_coverage_plot_for_region_in_domain(
  target_cells_lookup,
  cell_orog_lookup,
  region,
  domain,
  bg_map_sf = NULL,
  show_cell_labels = FALSE
)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{region}{Region to plot (either index in the list or region code, e.g. 37 or 'mudgee')}

\item{domain}{Domain to plot region and cells on}

\item{bg_map_sf}{An optional simple feature object to plot as a background map}

\item{show_cell_labels}{logical. If TRUE then shows the labels for the cell numbers. Set to FALSE by default.}
}
\value{
ggplot object for the plot
}
\description{
This plot uses the matching cells for specified domain.
}
\examples{
target_cells_lookup <- readRDS('/path/to/lookup_file.rds')
cell_orog_lookup <- readRDS('/path/to/orog_file.rds')

draw_spatial_coverage_plot_for_region_in_domain(target_cells_lookup, cell_orog_lookup, region = 'glenrowan', domain = 'VIC-5')
draw_spatial_coverage_plot_for_region_in_domain(target_cells_lookup, cell_orog_lookup, region = 'glenrowan', domain = 'SEA-10')
}
