% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{draw_spatial_coverage_plot_for_region_in_domain_with_cell_numbers}
\alias{draw_spatial_coverage_plot_for_region_in_domain_with_cell_numbers}
\title{Draw a spatial coverage plot for a region on a specific domain with the specified cell numbers}
\usage{
draw_spatial_coverage_plot_for_region_in_domain_with_cell_numbers(
  target_cells_lookup,
  cell_orog_lookup,
  region,
  domain,
  target_cells,
  show_cell_labels = FALSE
)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{region}{Region to plot (either index in the list or region code, e.g. 37 or 'mudgee')}

\item{domain}{Domain to plot region and cells on}

\item{target_cells}{Vector of cells in the domain to draw on the plot of region}

\item{show_cell_labels}{logical. If TRUE then shows the labels for the cell numbers. Set to FALSE by default.}
}
\value{
ggplot object for the plot
}
\description{
This plot uses the specified cells for the specified domain
}
\examples{
target_cells_lookup <- readRDS('/path/to/lookup_file.rds')
cell_orog_lookup <- readRDS('/path/to/orog_file.rds')

all_cells <- lookup_cells_for_region_with_domain(target_cells_lookup, 'glenrowan', 'VIC-5')
not_all_cells <- all_cells[1:(length(all_cells) \%/\% 2)]

draw_spatial_coverage_plot_for_region_in_domain_with_cell_numbers(target_cells_lookup, cell_orog_lookup, region = 'glenrowan', domain = 'VIC-5', target_cells = not_all_cells, show_cell_labels = TRUE)
}
