% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure a directory exists}
\usage{
ensure_dir_exists(path, silent = FALSE)
}
\arguments{
\item{path}{The directory}

\item{silent}{Do not return the directory}
}
\value{
The directory
}
\description{
Create a directory and any required parent directories if it does not already exist.
}
\details{
Useful to do at the start of a script to ensure the directory exists later to save outputs to.
This is a convenience wrapper for dir.create(), with a small addition: the directory is returned by the function.
This allows the directory to be assigned to a variable on the same line as calling the function (see examples).
}
\examples{
\donttest{
output_directory <- ensure_dir_exists('/mnt/projects/my_project')
output_directory <- '/mnt/projects/my_project' \%>\% ensure_dir_exists
output_directory <- file.path('/mnt/projects/my_project', 'v1') \%>\% ensure_dir_exists

file_to_save <- '/path/to/file_to_save.rds'
ensure_dir_exists(dirname(file_to_save))
saveRDS(mtcars, file_to_save)
}
}
