% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{extract_and_calc}
\alias{extract_and_calc}
\title{Extract data from NetCDF files for all regions in a target cells lookup for given model/domain/variables (and optionally perform post-calculations)}
\usage{
extract_and_calc(
  target_inputs_table,
  target_cells_lookup,
  extract_directory,
  batch_size = 100,
  batch_limit = NA,
  batch_start_number = 1,
  post_process_function = NULL,
  post_process_params = NULL
)
}
\arguments{
\item{target_inputs_table}{A dataframe with list of input files to extract data from. Must contains the columns: model, domain, variable, file.
The model, domain and variable columns must correspond with the filename, and the variable column must correspond with name of the variable in the NetCDF file.}

\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{extract_directory}{Directory to output the extract}

\item{batch_size}{Number of cells in each batch. Each batch is saved to disk for later use. If this number is too high then performance will suffer. Set to 100 by default.}

\item{batch_limit}{Number of batches to process. Useful for testing script behaviour prior to running on a complete data set.}

\item{batch_start_number}{Begin the extract from a particular batch number. This allows the process to resume after a system crash, etc. Note: does not work in conjunction with batch_limit (leave batch_limit set to NA to avoid strange behaviour).}

\item{post_process_function}{Name of a function to process the data through after it is extracted from the source files}

\item{post_process_params}{Parameters to pass to \code{post_process_function}}
}
\value{
Data frame with the extracted (and calculated) data
}
\description{
Extract data from NetCDF files for all regions in a target cells lookup for given model/domain/variables (and optionally perform post-calculations)
}
