% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{extract_cells_from_files}
\alias{extract_cells_from_files}
\alias{load_files_for_cells}
\title{Extract cell data from files}
\usage{
extract_cells_from_files(target_files, target_cells = NULL)

load_files_for_cells(target_files, target_cells = NULL)
}
\arguments{
\item{target_files}{Vector of files}

\item{target_cells}{Vector of cells to extract data for (if NULL, then all cells will be extracted)}
}
\value{
Dataframe
}
\description{
Data for only given cells will be extracted from the list of files
}
\details{
This is designed to be used in conjunction with \code{identify_files_for_target_cells},
and simplifies the process for extracting data from files where the cells of interest are spread across multiple files.
If a file does not have any of the cells in \code{target_cells} then no cells from the file will be added
(however, there will of course be an unnecessary increase in computation time as the file is loaded for no reason).

A warning message is shown if some \code{target_cells} are missing from the final data frame.
}
