% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{file_doesnt_exist_or_force}
\alias{file_doesnt_exist_or_force}
\title{Run if file doesn't exist or if force flag is set}
\usage{
file_doesnt_exist_or_force(file, force = FALSE)
}
\arguments{
\item{file}{Filename}

\item{force}{If TRUE, then will return TRUE regardless of whether the file exists or not}
}
\value{
Logical
}
\description{
Run if file doesn't exist or if force flag is set
}
\examples{
\donttest{
cache_file <- '/path/to/file.rds'
force_calc <- FALSE

if(file_doesnt_exist_or_force(cache_file, force_calc)) {
  # do something to create 'dataframe'
  saveRDS(dataframe, cache_file)
} else {
  dataframe <- readRDS(cache_file)
}
}
}
