% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{find_closest_cell_to_lat_lon_values}
\alias{find_closest_cell_to_lat_lon_values}
\title{Find closest cell to given lat/lon values}
\usage{
find_closest_cell_to_lat_lon_values(cell_orog_lookup, lat, lon, domain = NULL)
}
\arguments{
\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{lat}{Latitude, in decimal degrees}

\item{lon}{Longitude, in decimal degrees}

\item{domain}{Domain (not required if cell orography lookup has only one domain)}
}
\value{
Cell number
}
\description{
Find the cell in a domain that is closest to the given latitute and longitude.
This is useful where it cannot be relied upon that the lat/lon values directly correspond to a cell in the given domain.
}
\details{
The geosphere::distHaversine() function is used to detemine the distance between the lat/lon values and the centre of each cell.
}
\examples{
\donttest{
cell_orog_lookup <- readRDS('/path/to/file.rds')
cell <- get_closest_cell_to_lat_lon_values(cell_orog_lookup,
                                           lat = -42.8824,
                                           lon = 147.3301,
                                           domain = 'name_of_domain')
}
}
