% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{find_matching_cells_for_polygon_from_spatial_objects}
\alias{find_matching_cells_for_polygon_from_spatial_objects}
\title{Find the cells in a spatial array covered by a spatial object}
\usage{
find_matching_cells_for_polygon_from_spatial_objects(
  geospatial_feature_spdf,
  geospatial_array
)
}
\arguments{
\item{geospatial_feature_spdf}{SpatialPolygons object}

\item{geospatial_array}{Raster\* object (or a SpatialPixels\* or SpatialGrid\* object)}
}
\value{
Vector of cells
}
\description{
The geospatial feature
The geospatial array can be produced from a raster of a NetCDF file.
}
\details{
It is important that both parameters have the same projection.

TODO:
explain the usage of cellFromPolygon with weights = TRUE and why that is important
allow a cutoff to be passed for the weights
}
\examples{
shape_file <- '/path/to/shapes/region.shp'
array_file <- '/path/to/source/data/tasmax_monthly_1960-1969.nc'
matching_cells <- find_matching_cells_for_polygon_from_spatial_objects(shape_file, array_file)
}
