% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{generate_batch_end_vector}
\alias{generate_batch_end_vector}
\title{Generate vector of ending cells for batch processing}
\usage{
generate_batch_end_vector(batch_size, total_cells)
}
\arguments{
\item{batch_size}{Size of each batch}

\item{total_cells}{Total number of cells in extract}
}
\value{
Vector of ending cells
}
\description{
Generate vector of ending cells for batch processing
}
\examples{
generate_batch_end_vector(10, 55)  # 10, 20, 30, 40, 55
}
