% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{generate_batch_start_vector}
\alias{generate_batch_start_vector}
\title{Generate vector of starting cells for batch processing}
\usage{
generate_batch_start_vector(batch_size, total_cells)
}
\arguments{
\item{batch_size}{Size of each batch}

\item{total_cells}{Total number of cells in extract}
}
\value{
Vector of starting cells
}
\description{
Generate vector of starting cells for batch processing
}
\examples{
cfdtk:::generate_batch_start_vector(10, 55)
}
