% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{generate_output_filename_prefix}
\alias{generate_output_filename_prefix}
\title{Generate filename for CCAM extract}
\usage{
generate_output_filename_prefix(
  target_model,
  target_domain,
  cell_start,
  cell_end,
  last_cell
)
}
\arguments{
\item{target_model}{Model that data is for}

\item{target_domain}{Domain that data is for}

\item{cell_start}{The first cell in this data file}

\item{cell_end}{The last cell in this data file}

\item{last_cell}{The last cell for the entire variable, to determine number padding for every file}
}
\value{
String to be used at the filename
}
\description{
Format is \code{model-name_domain_cell-start_cell-end} where each component has no underscores
}
