% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{generate_target_cells_lookup_entry}
\alias{generate_target_cells_lookup_entry}
\title{Return list data structure for a target cells lookup entry}
\usage{
generate_target_cells_lookup_entry(
  id,
  code,
  label,
  sf,
  domain,
  matching_cells,
  matching_cells_with_weights
)
}
\value{
List
}
\description{
Return list data structure for a target cells lookup entry
}
