% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_all_sf_objects_from_target_cells_lookup}
\alias{get_all_sf_objects_from_target_cells_lookup}
\title{Get all of the embedded sf objects from a target cells lookup}
\usage{
get_all_sf_objects_from_target_cells_lookup(
  target_cells_lookup,
  proj4string = common_defaults$proj4string
)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}
}
\value{
An sfc object
}
\description{
Get all of the embedded sf objects from a target cells lookup
}
