% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_cells_from_orog_lookup}
\alias{get_cells_from_orog_lookup}
\alias{get_cells}
\title{Get cell orography data for specific cells}
\usage{
get_cells_from_orog_lookup(cell_orog_lookup, target_cells, domain)

get_cells(cell_orog_lookup, target_cells, domain)
}
\arguments{
\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{target_cells}{Vector of cells to retrieve data for}

\item{domain}{Domain to retrieve data for}
}
\value{
Data frame with cell orography data
}
\description{
Get the cell orography data for a specific set of cell numbers from the orography lookup for the chosen domain
}
\examples{
cell_orog_lookup <- readRDS('/path/to/file.rds')
target_cells <- c(987, 1597, 2584, 4181, 6765)
domain <- 'AUS-50'
get_cells_from_orog_lookup(cell_orog_lookup, target_cells, domain)
}
