% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_cells_from_orog_lookup_for_region_domain}
\alias{get_cells_from_orog_lookup_for_region_domain}
\alias{get_cells_for_region_in_domain}
\title{Get cell orography data for region}
\usage{
get_cells_from_orog_lookup_for_region_domain(
  target_cells_lookup,
  cell_orog_lookup,
  region,
  domain
)

get_cells_for_region_in_domain(
  target_cells_lookup,
  cell_orog_lookup,
  region,
  domain
)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{region}{Region. Can be the numerical index or the name of the list item in \code{target_cells_lookup} code}

\item{domain}{Domain}
}
\value{
Vector of cells
}
\description{
Get the cell orography data for the matching cells of the region from the orography lookup for the chosen domain
}
\examples{
target_cells_lookup <- readRDS('/path/to/target_cells_lookup_file.rds')
cell_orog_lookup <- readRDS('/path/to/cell_orog_lookup_file.rds')
region <- 'region_name'
domain <- 'AUS-50'
get_cells_from_orog_lookup_for_region_domain(target_cells_lookup, cell_orog_lookup, region, domain)
}
