% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_closest_cell_to_lat_lon_values}
\alias{get_closest_cell_to_lat_lon_values}
\title{Get the closest cell to given lat/lon values}
\usage{
get_closest_cell_to_lat_lon_values(cell_orog_lookup, lat, lon, domain = NULL)
}
\arguments{
\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{lat}{Latitude, in decimal degrees}

\item{lon}{Longitude, in decimal degrees}

\item{domain}{Domain (not required if cell orography lookup has only one domain)}
}
\value{
Cell number
}
\description{
Find the cell in a domain is closest to the given latitute and longitude, where the lat/lon does not have to be in the given domain.
}
\details{
The geosphere::distHaversine() function is used to detemine the distance between the lat/lon values and the centre of each cell.
}
