% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_scale_limits_from_values}
\alias{get_scale_limits_from_values}
\title{Title}
\usage{
get_scale_limits_from_values(
  values,
  interval = 1,
  snap_lower_to_zero = TRUE,
  snap_upper_to_zero = TRUE,
  symmetrical = FALSE
)
}
\arguments{
\item{symmetrical}{Make scale symmetrical over zero

Behaviour will be strange with symmetrical and snap set}
}
\value{

}
\description{
tests to write:
}
\examples{

}
