% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{headtail}
\alias{headtail}
\title{Return the First and Last Parts of an Object}
\usage{
headtail(x, n = 6L, ...)
}
\arguments{
\item{x}{an object}

\item{n}{a single integer. If positive, size for the resulting
    object: number of elements for a vector (including lists), rows for
    a matrix or data frame or lines for a function. If negative, all but
    the \code{n} last/first number of elements of \code{x}.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Combined output of \code{head} and \code{tail}
}
\description{
Return the First and Last Parts of an Object
}
\examples{
headtail(datasets::iris)
}
