% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{hist2d}
\alias{hist2d}
\title{Categorise two-dimensional data into bins (2D histogram)}
\usage{
hist2d(xy, x_breaks, y_breaks)
}
\arguments{
\item{xy}{Matrix with 2 columns}

\item{x_breaks}{Vector with breaks for the x column of the matrix}

\item{y_breaks}{Vector with breaks for the y column of the matrix}
}
\description{
Categorise two-dimensional data into bins (2D histogram)
}
\examples{
xy <- iris \%>\% dplyr::select(x = Sepal.Length, y = Sepal.Width)
x_breaks <- seq(from = plyr::round_any(min(xy$x), 0.5, floor),
                to = plyr::round_any(max(xy$x), 0.5, ceiling),
                by = 0.5)
y_breaks <- seq(from = plyr::round_any(min(xy$y), 0.5, floor),
                to = plyr::round_any(max(xy$y), 0.5, ceiling),
                by = 0.5)
hist2d(xy, x_breaks, y_breaks)
}
