% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{identify_files_for_target_cells}
\alias{identify_files_for_target_cells}
\title{Identify Files with Data for Specified Cells}
\usage{
identify_files_for_target_cells(
  extract_directory,
  target_model,
  target_domain,
  target_cells
)
}
\arguments{
\item{extract_directory}{Directory with saved data files with the containing cells in the filename, e.g. "0365-1389"}

\item{target_cells}{Vector of cells}
}
\value{
Vector of files
}
\description{
Identify Files with Data for Specified Cells
}
