% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{identify_files_for_target_model_input_thredds}
\alias{identify_files_for_target_model_input_thredds}
\title{Identify input files from data archive for model/domain/variable combinations}
\usage{
identify_files_for_target_model_input_thredds(
  target_models,
  target_domains,
  target_variables,
  source_directory = "/rdsi/private/climatefutures/WINE/output",
  org = "CSIRO",
  data_group = "rcp85",
  model_initial_conditions = "r1i1p1",
  model_version = "CSIRO-CCAM-r3355",
  data_version = "v1",
  tr = "day"
)
}
\arguments{
\item{target_models}{Vector of models (corresponding to directory structure)}

\item{target_domains}{Vector of domains (corresponding to directory structure)}

\item{target_variables}{Vector of variables of interest (corresponding to directory structure)}

\item{source_directory}{Source data files for climate data}

\item{org}{Thredds: organisation}

\item{data_group}{Thredds: data group}

\item{model_initial_conditions}{Thredds: model initial conditions}

\item{model_version}{Thredds: model version}

\item{data_version}{Thredds: data version}

\item{tr}{Thredds: temporal resolution}
}
\value{
A dataframe with columns with the model, domain, variable and filename of each file found
}
\description{
This follows the THREDDS conventions of directory structure and file naming
}
\details{
Create a data frame where each row represents a file to extract, and the corresponding model/domain/variable.
Uses expand.grid on a list of models, domains and variables; then searches the source_directory for files matching
those in the directory structure convention used by the THREDDS data server.
}
