% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{identify_files_with_data_for_target_cells}
\alias{identify_files_with_data_for_target_cells}
\title{Identify files with data for specified cells}
\usage{
identify_files_with_data_for_target_cells(
  extract_data_dir,
  target_model,
  target_domain,
  target_cells
)
}
\arguments{
\item{extract_data_dir}{Directory with saved data files with the containing cells in the filename, e.g. "0365-1389"}

\item{target_model}{Model}

\item{target_domain}{Domain}

\item{target_cells}{Vector of cells}
}
\value{
Vector of files
}
\description{
Determines which files in a directory of batches contain the data for a list of cells
}
