% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check if year is a leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{Year or vector of to test for being a leap year}
}
\value{
TRUE or FALSE
}
\description{
The calculation is derived from the article here: http://en.wikipedia.org/wiki/Leap_year
}
\examples{
is_leap_year(1999)  # FALSE
is_leap_year(2000)  # TRUE
is_leap_year(2100)  # FALSE
}
