% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{is_zero_range_vector}
\alias{is_zero_range_vector}
\title{Check if vector has range of zero (i.e. all elements the same)}
\usage{
is_zero_range_vector(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{Vector}

\item{tol}{Tolerance, defaults to .Machine$double.eps (for high precision with doubles)}
}
\value{
TRUE or FALSE
}
\description{
Checks if range of vector is zero, , or length of vector is one
}
\details{
As implemented here: https://stackoverflow.com/a/4752580
}
\examples{
is_zero_range_vector(c(1, 2, 3))
is_zero_range_vector(c(3.14, 3.14, 3.14))
}
