% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{isFALSE}
\alias{isFALSE}
\title{Check for FALSE}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{any object}
}
\value{
returns TRUE if x is identical to FALSE, FALSE otherwise.
}
\description{
Shortcut for identical(x, FALSE)
}
\details{
ggnewscale and other packages use the built in function isFALSE (for R >= 3.5).
This is a replacement, until R is upgraded to >= 3.5
}
\examples{
isFALSE(0)
isFALSE(FALSE)
}
