% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{label_to_code}
\alias{label_to_code}
\title{Convert a "label" string to a "code" string}
\usage{
label_to_code(label)
}
\arguments{
\item{label}{Label string to convert}
}
\value{
String
}
\description{
String will be converted to lowercase with underscores in place of spaces, and punctuation removed
}
\examples{
label_to_code('Tasmania East Coast')  # tasmania_east_coast
label_to_code("This/That won't be the Other")  # this_that_wont_be_the_other
label_to_code('Description (Extra Info)')  # description_extra_info
label_to_code('Description - Extra Info')  # description_extra_info
}
