% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{last_date_vector_under_threshold}
\alias{last_date_vector_under_threshold}
\title{Identify last date in a vector that an associated vector of values is under a given threshold}
\usage{
last_date_vector_under_threshold(dates, values, threshold)
}
\arguments{
\item{dates}{Vector of dates}

\item{values}{Vector of values}

\item{threshold}{Threshold value to test against}
}
\value{
The item in the dates vector which corresponds to the item in the values vector.
If the values vector never drops below the threshold, then NA is returned.
}
\description{
For a vector of dates and a corresponding vector of values, finds the last date which is less than a particular threshold value
}
\examples{
df <- data.frame(date = seq(as.Date("1961-07-01"), as.Date("1961-07-31"), "days"),
                 vals = seq(from = 31, to = 1, by = 1) )
last_date_vector_under_threshold(df$date, df$vals, 4)
}
