% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{logpipe}
\alias{logpipe}
\title{Helper log function to monitor progress of a dplyr pipe chain}
\usage{
logpipe(x, msg)
}
\arguments{
\item{x}{The data frame being processed (so that it can be returned unaltered)}

\item{msg}{Message to log}
}
\description{
Helper log function to monitor progress of a dplyr pipe chain
}
\examples{
\donttest{
mtcars \%>\%
  logpipe('Test 1') \%>\%
  group_by(mpg) \%>\%
  {Sys.sleep(2); .} \%>\%
  logpipe('Test 2') \%>\%
  tally()
}
}
