% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{logtime}
\alias{logtime}
\alias{logtime_print_diff}
\alias{logtime_secs_diff}
\alias{logtime_init}
\title{Log and display the time taken for tasks to complete}
\usage{
logtime(time_log, log_name, log_type = NULL, .start = FALSE, .end = FALSE)

logtime_print_diff(time_log, log_name, digits = 4)

logtime_secs_diff(time_log, log_name)

logtime_init()
}
\arguments{
\item{time_log}{An existing logtime object}

\item{log_name}{Name of the log entry}

\item{log_type}{State of the log entry, either 'start' or 'end'}

\item{.start}{Alternative method of defining the log type; set to TRUE for a 'start' log type}

\item{.end}{Alternative method of defining the log type; set to TRUE for an 'end' log type}

\item{digits}{How many significant digits are to be used for printed value}
}
\description{
Provides a set of tools to allow easier reporting on the time taken to perform intermediate tasks within a larger body of code.
}
\details{
First, initialise new logtime object with the \code{logtime_init} method.
Then, before starting a task to measure the time taken,
use \code{logtime} to append a 'start' timestamp for a given name (the \code{log_name}) to refer uniquely to this task.
After the task has completed,
use \code{logtime} again to append an 'end' timestamp for the same \code{log_name}.

To print the time difference between the start and end times of a log entry, use the \code{logtime_print_diff} method.
This will use the most suitable measurement of time to represent the elapsed time to make the entry readable.

To retrieve the time difference between the start and end times of a log entry in seconds, use the \code{logtime_secs_diff} method.
This is sdesigned to be used for tests on the time difference.
}
\examples{
time_log <- logtime_init()
time_log <- logtime(time_log, 'my_log_name', .start = TRUE)
# do some things
# that take time
time_log <- logtime(time_log, 'my_log_name', .end = TRUE)
logtime_print_diff(time_log, 'my_log_name')
}
