% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{lookup_cells_for_domain}
\alias{lookup_cells_for_domain}
\alias{get_cells_for_domain}
\title{Get target cell indices for a domain from lookup object}
\usage{
lookup_cells_for_domain(target_cells_lookup, target_domain)

get_cells_for_domain(target_cells_lookup, target_domain)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{target_domain}{The domain in the lookup object to get the matching cells for}
}
\value{
An integer vector of unique cells, not necessarily in sequencial order
}
\description{
Return the cells in the specified domain that cover all of the regions.
Duplicate cells (i.e. cells that exist in multiple regions) will be listed only once.
}
