% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{get_cells_for_region}
\alias{get_cells_for_region}
\alias{lookup_cells_for_region}
\title{Lookup cells for a region}
\usage{
get_cells_for_region(cell_orog_lookup, target_cells_lookup, region)

lookup_cells_for_region(target_cells_lookup, region)
}
\arguments{
\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{region}{Can be the numerical index or the name of the list element}
}
\value{
Vector of cells for \code{lookup_*} functions, tibble with cell orography data for \code{get_*} functions
}
\description{
Retrieve the cells for a region, in the domain specified to have the highest resolution data
}
