% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{lookup_domain_for_region}
\alias{lookup_domain_for_region}
\title{Lookup preferred domain for a region}
\usage{
lookup_domain_for_region(target_cells_lookup, region)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{region}{Can be the numerical index or the name of the list element}
}
\value{
Domain
}
\description{
Retrieve the domain specified to have the highest resolution data for a region
}
