% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{lookup_polygon_label_from_code}
\alias{lookup_polygon_label_from_code}
\title{Lookup region label from the code}
\usage{
lookup_polygon_label_from_code(target_cells_lookup, region)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{region}{Region code}
}
\value{
Label
}
\description{
Lookup region label from the code
}
