% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{pick_preferred_domain_from_lookup}
\alias{pick_preferred_domain_from_lookup}
\title{Pick the preferred domain for a polygon when building a target cells lookup}
\usage{
pick_preferred_domain_from_lookup(
  preferred_domains,
  target_domains_files,
  polygon_code
)
}
\arguments{
\item{preferred_domains}{Optional. A named list defining the preferred domain for each polygon.
If the polygons are spread across multiple domains, then the preferred domain for each polygon must be specified explictly to distinguish between those that overlap.
Use the code for the polygon (as will be created in \code{convert_all_polygons_to_multipolygons}) as the key, and the doman as the value.
If only one domain is present in the \code{target_domains_files} parameter, then this list is not required.}

\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}

\item{polygon_code}{}
}
\value{
Domain to be chosen
}
\description{
If there is only one domain, then set the preferred domain as that.
Otherwise, a list of which domain is the preferred to list the matching cells must be used.
}
