% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{progress_of_loop}
\alias{progress_of_loop}
\alias{loginfo_progress}
\title{Build and/or log string describing current progress of a loop}
\usage{
progress_of_loop(element, all_elements, FUN = as.character)

loginfo_progress(element, all_elements, n = 1, FUN = as.character)
}
\arguments{
\item{element}{Current element of the loop}

\item{all_elements}{All elements that will be iterated over}

\item{FUN}{Function to pass element to before logging. For example, \code{basename} will shorten the log line for a file with a long path that is assumed or logged previously.}

\item{n}{Every \code{n}th element will be logged. The first and last element are always logged. Useful for loops with a large number of iterations.}
}
\value{
String
}
\description{
Helper function to build a string to be logged that describes the progress within a loop (e.g. for, lapply, ...).
}
\details{
\code{progress_of_loop} will return a string matching the format "(x/y) element" will be returned,
where 'x' is the index of the current element in the list,
'y' is the total number of elements,
and 'element' is the element itself.

\code{loginfo_progress} will log a string as produced by \code{progress_of_loop} to the info logger.
}
\examples{
for (letter in letters) {
  loginfo(progress_of_loop(letter, letters))
  # do something
}

lapply(letters, function(letter)  {
  loginfo(progress_of_loop(letter, letters))
  stringr::str_to_upper(letter)
}) \%>\% unlist
\donttest{
all_files <- glue("/path/to/file/{letters}.txt")

loginfo_progress(all_files[12], all_files)

loginfo_progress(all_files[1], all_files, FUN = basename)
loginfo_progress(all_files[5], all_files, FUN = basename)
}
}
