% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{QuickView}
\alias{QuickView}
\alias{QV}
\title{Quickly view a large object in the data viewer}
\usage{
QuickView(x, n = 50, title = NULL)

QV(x, n = 50, title = NULL)
}
\arguments{
\item{x}{an \R object which can be coerced to a data frame with
    non-zero numbers of rows and columns.}

\item{n}{Number of rows to show in data viewer}

\item{title}{title for viewer window.  Defaults to name of \code{x}
    prefixed by \code{Data:}.}
}
\description{
By default the first 50 rows of an object will be shown.
}
\details{
This is intended as a quick way to type this in the console,
by typing the name of a dataframe and piping it to \code{QV} using the RStudio keyboard shortcut Ctrl+Shift+m.
}
\examples{
QuickView(iris)
QV(iris)
}
