% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{save_png}
\alias{save_png}
\alias{save_png_a4}
\title{Save a grob to a PNG file}
\usage{
save_png(
  grob,
  file,
  width = 1920,
  height = 1080,
  units = "px",
  res = "144",
  silent = FALSE
)

save_png_a4(
  grob,
  file,
  width = NULL,
  height = NULL,
  units = "mm",
  res = 200,
  silent = FALSE,
  preserve_filename = FALSE
)
}
\arguments{
\item{grob}{grob of plot to save}

\item{file}{the name of the output file}

\item{width}{the width of the image}

\item{height}{the height of the image}

\item{units}{the units in which \code{width} and \code{height} are given}

\item{res}{resolution of the image in ppi}

\item{silent}{Default: FALSE. If TRUE, will not display a loginfo message.}

\item{preserve_filename}{Preserve the original filename (i.e. do not prefix with "A4_")}
}
\description{
Files are saved using some sensible defaults for general analysis usage.
}
\details{
Default dimensions are 1920x1080 (to correspond with those of a standard large screen monitor). In addition, the png device will still be closed even if there is an error with the plot. This means that the file will still be created (with an empty canvas), and no lingering device will be left open.

\code{save_png_a4} sets the dimension parameters to be such that the width takes an A4 page minus 25.4mm margins and the width:height ratio of the image is 4:3
(meaning the default width is 159.2mm and the default height is 119.4mm).
The text theme \code{grob} will be automatically be scaled to a text size of 9.5, in order to make the text size slightly smaller than standard 12pt text in a document.
In addition, the filename will be prefixed with "A4_" (if not already) to make it easier to differentiate between the A4 image and the screen image.
This automatic prefix can be prevented by setting the \code{preserve} parameter to \code{TRUE} (it is \code{FALSE} by default).
}
\examples{
\donttest{
grob <- ggplot() + geom_point(data = datasets::mtcars, mapping = aes(x = cyl, y = hp))
file <- '/path/to/output/figures/img.png'
save_png(grob, file)
}
}
