% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{save_rds}
\alias{save_rds}
\title{Save data frame to rds file, adding or changing the extension of \code{file} if required}
\usage{
save_rds(x, file)
}
\arguments{
\item{x}{Data frame}

\item{file}{Filename}
}
\description{
Save data frame to rds file, adding or changing the extension of \code{file} if required
}
