% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{scale_x_datetime_decadal}
\alias{scale_x_datetime_decadal}
\title{Title}
\usage{
scale_x_datetime_decadal(dates, ...)
}
\arguments{
\item{dates}{Vector of dates (POSIXct or date)}

\item{...}{Parameters passed to \code{scale_x_datetime}}
}
\description{
Title
}
