% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{seconds_in_month}
\alias{seconds_in_month}
\title{Get the number of seconds in the month of a date-time}
\usage{
seconds_in_month(x)
}
\arguments{
\item{x}{a date-time object}
}
\description{
Get the number of seconds in the month of a date-time
}
