% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{seconds_per_month}
\alias{seconds_per_month}
\title{Calculate number of seconds per month}
\usage{
seconds_per_month(year, month)
}
\arguments{
\item{year}{Year}

\item{month}{Month}
}
\value{
Integer
}
\description{
This will take into account leap years
}
\examples{
seconds_per_month(1979, 1)
seconds_per_month(1979, 2)
seconds_per_month(1979, 3)
seconds_per_month(1979, 4)
seconds_per_month(1980, 2)
}
