% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{select_files_for_target_domain_and_variables}
\alias{select_files_for_target_domain_and_variables}
\title{Select files to process from a pre-calculated list for target variables in a target domain}
\usage{
select_files_for_target_domain_and_variables(
  target_inputs_table,
  target_domain,
  target_variables
)
}
\arguments{
\item{target_inputs_table}{A dataframe with list of input files to extract data from. Must contains the columns: model, domain, variable, file.}

\item{target_domain}{Domain}

\item{target_variables}{A vector of variables}
}
\value{
A named list with variables as the keys, and the values being a vector of filenames
}
\description{
Select files to process from a pre-calculated list for target variables in a target domain
}
