% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{set_names_for_target_cells_lookup}
\alias{set_names_for_target_cells_lookup}
\title{Set the names of a target cells lookup list to the code property of each element}
\usage{
set_names_for_target_cells_lookup(target_cells_lookup)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}
}
\description{
Set the names of a target cells lookup list to the code property of each element
}
