% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{setup_common_project_directories}
\alias{setup_common_project_directories}
\title{Adds variables for common project directories to the global environment and ensures they exist}
\usage{
setup_common_project_directories(project_directory)
}
\arguments{
\item{project_directory}{Base project directory}
}
\description{
Adds variables for common project directories to the global environment and ensures they exist
}
