% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{target_cells_lookup_summary_by_number_of_cells}
\alias{target_cells_lookup_summary_by_number_of_cells}
\alias{target_cells_lookup_sort_by_number_of_cells}
\alias{target_cells_lookup_plot_by_number_of_cells}
\title{Display a simple summary of the number of cells for each region in the target cells lookup}
\usage{
target_cells_lookup_summary_by_number_of_cells(target_cells_lookup)

target_cells_lookup_sort_by_number_of_cells(target_cells_lookup)

target_cells_lookup_plot_by_number_of_cells(target_cells_lookup)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}
}
\value{
Dataframe with columns: region, domain, number of cells

A sorted lookup

ggplot object of the plot
}
\description{
Returns the target cells lookup itself, sorted in the order of least cells to most cells.
}
\details{
This is useful when performing long batch processes, as smaller regions will be processed first,
meaning the results for a greater number of regions will be ready to to inspect sooner
}
