% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{vector_to_c}
\alias{vector_to_c}
\title{Takes an R vector and prints a string that can be used to define that vector}
\usage{
vector_to_c(x)
}
\arguments{
\item{x}{Vector}
}
\value{
A string (glue object) that is printed to the console
}
\description{
Takes an R vector and prints a string that can be used to define that vector
}
\examples{
month.abb
vector_to_c(month.abb)
}
