% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{verify_batch_files}
\alias{verify_batch_files}
\title{Verify all batch files are present}
\usage{
verify_batch_files(
  extract_directory,
  target_cells_lookup,
  target_domain,
  target_model,
  batch_size
)
}
\arguments{
\item{extract_directory}{Directory to output the extract}

\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{target_domain}{Domain that data is for}

\item{target_model}{Model that data is for}

\item{batch_size}{Number of cells in each batch. Each batch is saved to disk for later use. If this number is too high then performance will suffer. Set to 100 by default.}
}
\value{
A data frame with a row for each batch specifying the batch number,
the first and last cell in the batch, the expected filename pattern,
and the name of the actual file (if it exists)
}
\description{
Examine an extract directory and verify that each file is named as expected given a set of batch parameters.
}
\details{
This is useful if an extract was performed in multiple stages for resourcing reasons, to check that each batch has been processed correctly.
}
