% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{verify_polygon_coverage_for_extracted_cells}
\alias{verify_polygon_coverage_for_extracted_cells}
\title{Verify if all the cells for a region have been extracted}
\usage{
verify_polygon_coverage_for_extracted_cells(
  extract_directory,
  target_cells_lookup,
  target_region_code
)
}
\arguments{
\item{extract_directory}{Directory with cell extract files}

\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{target_region_code}{Region code}
}
\value{
List with attributes describing the cells for each region, the cells in the extract, whether all cells have been extracted, and the number still to be extracted.
}
\description{
Useful for beginning work on an extract in process
}
