% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{write_to_file}
\alias{write_to_file}
\alias{write_to_csv_file}
\alias{write_to_RData_file}
\alias{write_to_RDS_file}
\title{Writes calculated stats object to file on disk}
\usage{
write_to_file(x, output_directory, filename, format)

write_to_csv_file(x, output_directory, filename)

write_to_RData_file(x, output_directory, filename)

write_to_RDS_file(x, output_directory, filename)
}
\arguments{
\item{x}{Object to save}

\item{output_directory}{Directory to save file to}

\item{filename}{Name of the file (without the extension)}

\item{format}{File format to save; one of: 'csv', 'RData', 'rds'}
}
\description{
This collection of wrapper functions handle a number of general tasks, such as ensuring the target directory exists and logging the start/end process and the size of the file.
}
\examples{
\donttest{
write_to_file(datasets::mtcars, '/path/to/dir', 'project_name', 'csv')
}
}
