% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{common_parameter_descriptions}
\alias{common_parameter_descriptions}
\title{Common parameter descriptions}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}

\item{lat}{Latitude, in decimal degrees}

\item{lon}{Longitude, in decimal degrees}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}

\item{field_names}{A named list with the following names: 'index', 'label', and 'code'}

\item{preferred_domains}{Optional. A named list defining the preferred domain for each polygon.
If the polygons are spread across multiple domains, then the preferred domain for each polygon must be specified explictly to distinguish between those that overlap.
Use the code for the polygon (as will be created in \code{convert_all_polygons_to_multipolygons}) as the key, and the doman as the value.
If only one domain is present in the \code{target_domains_files} parameter, then this list is not required.}
}
\description{
A placeholder to contain descriptions for parameters that are used in multiple functions
}
\keyword{internal}
