% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_date_breaks_decadal}
\alias{create_date_breaks_decadal}
\title{Title}
\usage{
create_date_breaks_decadal(dates)
}
\arguments{
\item{dates}{Vector of dates (POSIXct or date)}
}
\value{

}
\description{
for scale_x_datetime
}
