% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_or_read}
\alias{create_or_read}
\title{Load pre-calculated results from file if it exists, if not calculate results as save it to the file}
\usage{
create_or_read(file, FUN, ARGS, force = FALSE)
}
\arguments{
\item{file}{Name of the file}

\item{FUN}{Function (or code block) to create the object that is to be stored and subsequently retrieved from the \code{file}}

\item{ARGS}{List of arguments to be passed to the function}

\item{force}{logical. If TRUE, then force recreation and saving of the object, even if file already exists. Set to FALSE by default.}
}
\value{

}
\description{
Load pre-calculated results from file if it exists, if not calculate results as save it to the file
}
