% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{create_polygons_target_cells_lookup_from_points}
\alias{create_polygons_target_cells_lookup_from_points}
\title{Create a target cells lookup from a data frame of points with lat/lon values}
\usage{
create_polygons_target_cells_lookup_from_points(
  dataframe_locations,
  target_domains,
  cell_orog_lookup,
  target_domains_files,
  field_names = list(index = "id", label = "label", code = "code"),
  preferred_domains = list(),
  proj4string = common_defaults$proj4string
)
}
\arguments{
\item{dataframe_locations}{A data frame of locations representing points to extract.
Required columns: label, lat, lon.
Optional columns: index, code (these will be generated if they don't exist).}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}

\item{target_domains_files}{Named list defining each domain to be assessed with an example NetCDF file with the desired projection for each domain}

\item{field_names}{A named list with the following names: 'index', 'label', and 'code'}

\item{preferred_domains}{Optional. A named list defining the preferred domain for each polygon.
If the polygons are spread across multiple domains, then the preferred domain for each polygon must be specified explictly to distinguish between those that overlap.
Use the code for the polygon (as will be created in \code{convert_all_polygons_to_multipolygons}) as the key, and the doman as the value.
If only one domain is present in the \code{target_domains_files} parameter, then this list is not required.}

\item{proj4string}{A proj4string describing coordinate reference system to be used (set to "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0" by default}
}
\value{
A target cells lookup list
}
\description{
This lookup will specify each region/point of interest and which cells match each domain
}
\details{
From a set of lat/lon points of interest
}
