% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{date_vector_exceeds_threshold}
\alias{date_vector_exceeds_threshold}
\title{Identify date in a vector that an associated vector of values first exceeds a given threshold}
\usage{
date_vector_exceeds_threshold(dates, values, threshold)
}
\arguments{
\item{dates}{Vector of dates}

\item{values}{Vector of values}

\item{threshold}{Threshold value to test against}
}
\value{
The item in the dates vector which corresponds to the item in the values vector.
If the values vector never exceeds the threshold, then NA is returned.
}
\description{
For a vector of dates and a corresponding vector of values, finds the first date which is greater than a particular threshold value
}
\examples{
df <- data.frame(date = seq(as.Date("1961-07-01"), as.Date("1961-07-31"), "days"),
                 vals = c(1.1, 2.2, 3.3, 4.4, 5.5, 6.6, 7.7, 8.8, 9.9) )
date_vector_exceeds_threshold(df$date, df$vals, 5)
}
