% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{df_add_cell_orog_data}
\alias{df_add_cell_orog_data}
\title{Add cell orography information to a dataframe}
\usage{
df_add_cell_orog_data(df, cell_orog_lookup)
}
\arguments{
\item{df}{Dataframe}

\item{cell_orog_lookup}{Cell orography lookup, as created by \code{create_cell_orography_lookup}}
}
\value{

}
\description{
Add cell orography information to a dataframe
}
