% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{find_matching_cells_for_polygon_from_spatial_objects}
\alias{find_matching_cells_for_polygon_from_spatial_objects}
\title{Find the cells in a spatial array covered by a spatial object}
\usage{
find_matching_cells_for_polygon_from_spatial_objects(
  geospatial_feature_spdf,
  geospatial_array,
  weights_cutoff = 0
)
}
\arguments{
\item{geospatial_feature_spdf}{SpatialPolygonsDataFrame of the spatial object to be examined}

\item{geospatial_array}{Raster* object (or a SpatialPixels* or SpatialGrid* object)}

\item{weights_cutoff}{Cells with a weighting below this value will be excluded from the returned results}
}
\value{
Data frame with 2 columns: cell (the cell number), and weight (the weighting)
}
\description{
This functions uses \code{raster::cellFromPolygon} with \code{weights = TRUE} to determine which cells in the geospatial array are covered by the SpatialPolygonsDataFrame and what fraction of each cell is covered.
}
\details{
It is important that both the geospatial feature SpatialPolygonsDataFrame and the geospatial array have the same projection.
}
