% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{generate_target_cells_lookup_entry}
\alias{generate_target_cells_lookup_entry}
\title{Return list data structure for a target cells lookup entry}
\usage{
generate_target_cells_lookup_entry(
  id,
  code,
  label,
  sf,
  domain,
  matching_cells,
  matching_cells_with_weights
)
}
\arguments{
\item{id}{Value for id}

\item{code}{Value for code}

\item{label}{Value for label}

\item{sf}{An sf object for the polygon (for convenience when making plots)}

\item{domain}{Value for the preferred domain}

\item{matching_cells}{Named list for each domain with the cells as an integer vector}

\item{matching_cells_with_weights}{Named list for each domain with the cells and the percentage coverage they are in the polygon (weight) as a dataframe}
}
\value{
List
}
\description{
Return list data structure for a target cells lookup entry
}
