% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{label_to_code}
\alias{label_to_code}
\title{Convert a "label" string to a "code" string}
\usage{
label_to_code(label)
}
\arguments{
\item{label}{Label string to convert}
}
\value{
String
}
\description{
String will be converted to lowercase with underscores in place of spaces, and punctuation removed
}
\details{
There will never be more than one underscore in a row, and the code cannot start or end with an underscore
}
\examples{
label_to_code('Tasmania East Coast')  # tasmania_east_coast
label_to_code('Launceston (Tasmania)')  # launceston_tasmania
}
