% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{lookup_cells_for_region}
\alias{lookup_cells_for_region}
\alias{get_cells_for_region}
\title{Lookup cells for a region}
\usage{
lookup_cells_for_region(target_cells_lookup, region)

get_cells_for_region(target_cells_lookup, region)
}
\arguments{
\item{target_cells_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}

\item{region}{Can be the numerical index or the name of the list element}
}
\value{
Vector of cells
}
\description{
Retrieve the cells for a region, in the domain specified to have the highest resolution data
}
