% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{pick_preferred_domain_from_lookup}
\alias{pick_preferred_domain_from_lookup}
\title{Pick the preferred domain for a polygon when building a target cells lookup}
\usage{
pick_preferred_domain_from_lookup(
  preferred_domains,
  polygon_code,
  target_domains = NULL
)
}
\arguments{
\item{preferred_domains}{Optional. A named list defining the preferred domain for each polygon.
If the polygons are spread across multiple domains, then the preferred domain for each polygon must be specified explictly to distinguish between those that overlap.
Use the code for the polygon (as will be created in \code{convert_all_polygons_to_multipolygons}) as the key, and the doman as the value.
If only one domain is present in the \code{target_domains_files} parameter, then this list is not required.}

\item{polygon_code}{Polygon to get preferrend domain for}
}
\value{
Domain to be chosen
}
\description{
If there is only one domain in the target domain files, or if a single string is provided, then the preferred domain will be set as that.
Otherwise, a list of which domain is the preferred to list the matching cells must be used.
}
