% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfdtk.R
\name{read_part_of_filename}
\alias{read_part_of_filename}
\title{Read part from output of extract and calc}
\usage{
read_part_of_filename(file, n, pattern = "_")
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Read part from output of extract and calc
}
